backend/requirements.txt
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
configparser==7.1.0
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
executing==2.0.1
fastjsonschema==2.19.1

Flask==3.0.3
Flask_Cors==5.0.0
flask_jwt_extended==4.6.0
pymongo==4.10.1
python-dotenv==1.0.1
flask_mail==0.10.0
pycountry==24.6.1
pymongo==4.10.1
python-dotenv==1.0.1
Werkzeug==3.0.4
requests==2.32.3
flask_socketio==5.4.1
twilio==9.3.6
